Refactor resolution to param_resolution
