+++ src/multiview_stitcher/param_resolution.py
